utilities/analysis-report/requirements.txt
# Core data processing
pandas==2.0.3
numpy==1.24.3
openpyxl==3.1.2

# Database connectivity
psycopg2-binary==2.9.7

# Environment management
python-dotenv==1.0.0

# Utility libraries
tqdm==4.65.0

# Configuration management
PyYAML==6.0.1

# HTTP requests for filestore integration
requests==2.31.0
requests-toolbelt==1.0.0